Deepak-Product CRUD on progress

+++ src/main/java/miu/edu/cs545waa/repository/ProductCategoryRepo.java
@@ -6,5 +6,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProductCategoryRepo extends JpaRepository<ProductCategory, Long> {
+public interface ProductCategoryRepo extends JpaRepository<ProductCategory, Integer> {
 }
